Update project references and enhance error logging
The project references to SemanticKernel.Core in Connectors.AI.OpenAI.csproj were removed as they are longer needed. A ILogger was introduced for more effective logging in FunctionExtensions.cs. To catch unhandled exceptions in a cleaner way, error logging was added. There was also an update on how the JsonSerializer deserialize the functionResult in SKFunctionCallExtensions.cs to provide more accurate results.