Extend KernelArguments with additional methods
The KernelArguments class has been extended with new methods that allow for a more flexible management of arguments. This includes a method for initializing with an input string, updating the main input text, storing additional data, and printing out the processed data. The changes also include revisions to existing methods to align with these new functionalities.