.Net: Allow KernelFunction argument to KernelFunctionFromMethod method (#4119)
If a method wants to return additional metadata, it can do so via
returning a FunctionResult, but it can only construct a FunctionResult
if it has a reference to itself, which is currently only possible to get
in a convoluted manner. This allows the method to take a KernelFunction
argument that will be set to itself, just as it can take a Kernel,
KernelArguments, etc, argument to have access to that information.

(cherry picked from commit 7f1fe7e9583e61e0a2e96c123ffce703d5cf7d75)